consumer/requirements.txt
# pyspark
kafka-python